refactor: split function.h into exec_objects, class_objects, container_objects
function.h was a 250-line catch-all holding 10 unrelated object types across
three semantic domains. Split into focused headers:

- exec_objects.h: ObjFunction, ObjUpvalue, ObjClosure, ObjNative, NativeFn
  (absorbs native.h — which is deleted)
- class_objects.h: ObjClass, ObjInstance, ObjBoundMethod
- container_objects.h: ObjList, ObjIterator, ObjFile, ObjMap, ObjEnum/Ctor
- container_objects.cpp: ObjMap::mapGet/Set/Del, MapPolicy::hashOf
  (moved out of object.cpp)
- objects.h: umbrella include for consumers that need all types

Move isValueOfType<T> from function.h to value.h; rewrite the three
hand-coded isString/isFunction/isNative helpers as one-line delegations.
