+++ src/compiler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "function.h"
+#include "exec_objects.h"
 #include "parser.h"
 
 #include <cstdint>
