+++ src/math.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "native.h"
+#include "exec_objects.h"
 
 #include <cstddef>
