+++ src/objects.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "container_objects.h"
+#include "exec_objects.h"
+#include "class_objects.h"
